requirements.txt
absl-py==2.1.0
altair==5.4.1
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dlib<=19.24.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
idna==3.10
imageio==2.36.1
itsdangerous==2.2.0
jax==0.4.38
jaxlib==0.4.38
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml_dtypes==0.5.0
mpmath==1.3.0
narwhals==1.20.1
networkx==3.4.2
numpy==1.26.4
# opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pillow_heif==0.21.0
protobuf==4.25.5
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.14.1
sentencepiece==0.1.99
setuptools==75.7.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tifffile==2024.12.12
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.42.0
